# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/friends/status/{userId}": {
      "get": {
        "operationId": "FriendsController_getFriendshipStatus",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipStatusDto"
                }
              }
            }
          }
        },
        "summary": "Get friendship status with a specific user",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/request/{userId}": {
      "post": {
        "operationId": "FriendsController_sendFriendRequest",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipDto"
                }
              }
            }
          }
        },
        "summary": "Send a friend request to a user",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/accept/{id}": {
      "post": {
        "operationId": "FriendsController_acceptFriendRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipDto"
                }
              }
            }
          }
        },
        "summary": "Accept a friend request",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/decline/{id}": {
      "delete": {
        "operationId": "FriendsController_declineFriendRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Decline a friend request",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/cancel/{id}": {
      "delete": {
        "operationId": "FriendsController_cancelFriendRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Cancel a sent friend request",
        "tags": [
          "Friends"
        ]
      }
    },
    "/api/friends/{id}": {
      "delete": {
        "operationId": "FriendsController_removeFriend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Remove a friend (unfriend)",
        "tags": [
          "Friends"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kraken API",
    "description": "Kraken Chat Application API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "FriendshipStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true,
            "enum": [
              "PENDING",
              "ACCEPTED",
              "BLOCKED"
            ]
          },
          "direction": {
            "type": "string",
            "enum": [
              "sent",
              "received"
            ],
            "nullable": true
          },
          "friendshipId": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "status",
          "direction",
          "friendshipId"
        ]
      },
      "FriendshipDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "BLOCKED"
            ]
          },
          "id": {
            "type": "string"
          },
          "userAId": {
            "type": "string"
          },
          "userBId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "status",
          "id",
          "userAId",
          "userBId",
          "createdAt"
        ]
      },
...
    }
  }
}